Remove test
This tests fi this handles invalid body format, BUT, since the project is small and won't scale any way, I decided to KISS (Keep It Simple Stupid). So, there is no need to decouple server/client that much, if the client (that I programmed) is passing some bad format, this is my fault. Let this panic.